constantly_running/requirements.txt
boto3==1.14.9
botocore==1.17.9
certifi==2020.6.20
chardet==3.0.4
docutils==0.15.2
idna==2.9
jmespath==0.10.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
tzlocal==2.1
urllib3==1.25.9
